+++ spec/furi_spec.rb
@@ -95,6 +95,7 @@ describe Furi do
         query_tokens: [['a', '1'], ['b', '2']],
         query: {'a' => '1', 'b' => '2'},
         request: '/articles/index.html?a=1&b=2',
+        origin: 'http://user:pass@www.gusiev.com:8080/articles/index.html',
         anchor: 'header',
         fragment: 'header',
         home_page?: false,
@@ -115,6 +116,7 @@ describe Furi do
         resource: nil,
         resource!: '/',
         location: 'http://gusiev.com',
+        origin: 'http://gusiev.com',
         home_page?: true,
       )
     end
@@ -499,6 +501,10 @@ describe Furi do
       expect(Furi.replace("gusiev.com:433/index.html", location: 'gusiev.com:80')).to eq('gusiev.com:80/index.html')
       expect(Furi.replace("gusiev.com:433/index.html", location: nil)).to eq('/index.html')
       expect(Furi.replace("http://gusiev.com:433/index.html", location: nil)).to eq('/index.html')
+    end
+
+    it "replace origin" do
+      expect(Furi.replace("http://gusiev.com/index.html?a=1#top", origin: 'http://gusiev.com/blog.html')).to eq('http://gusiev.com/blog.html?a=1#top')
     end
 
     it "replace query" do